fixed jwt access-refresh token

+++ server/src/index.ts
@@ -1,127 +1,121 @@
 import express from 'express';
-import { connectDB } from './config/configDB';
-import { Content } from './model/content-schema';
-import { verifyuser } from './middleware/auth';
-import cookieParser from 'cookie-parser';
 import cors from 'cors';
+import { userMiddleware } from './middleware/auth';
 import { User } from './model/user-schema';
-import { JWT_ACCESS_PASSWORD, JWT_REFRESH_PASSWORD } from './config/configJWT';
-import jwt, { JwtPayload } from 'jsonwebtoken';
+import { Content } from './model/content-schema';
 import { Link } from './model/link-schema';
+import { Tags } from './model/tags-schema';
+import jwt, { JwtPayload } from "jsonwebtoken";
+import cookieParser = require("cookie-parser");
+import { connectDB } from './config/configDB';
+import { JWT_ACCESS_PASSWORD, JWT_REFRESH_PASSWORD } from './config/configJWT';
 import { random } from './lib/utils';
-const app = express();
 const port = 3000;
-
+const app = express();
 app.use(express.json());
-app.use(cookieParser());
 app.use(cors({
     origin: 'http://localhost:5173',
     credentials: true,
     allowedHeaders: ['Content-Type', 'Authorization', 'X-Requested-With'],
     exposedHeaders: ['Authorization']
 }));
+app.use(cookieParser());
 
-
-app.route('/api/v1/signup')
-    .post(async (req, res) => {
-        const username = req.body.username;
-        const password = req.body.password;
-
-        try {
-            await User.create({
-                username: username,
-                password: password
-            })
-
-            res.status(200).json({
-                message: "User signed up"
-            })
-        } catch (e) {
-            res.status(411).json({
-                message: "User already exists"
-            })
+app.post('/api/v1/signup', async (req, res) => {
+    const username = req.body.username;
+    const password = req.body.password;
+    try {
+        const existingUser = await User.findOne({ username });
+        if (existingUser) {
+            res.status(411).json({ message: 'User already exists' });
         }
-    });
+        await User.create({
+            username: username,
+            password: password
+        });
+        res.json({ message: 'User created successfully' });
+    } catch (error) {
+        res.status(411).json({ message: error });
+    }
+});
 
-app.route('/api/v1/signin')
-    .post(async (req, res) => {
-        const username = req.body.username;
-        const password = req.body.password;
-        const existinguser = await User.findOne({ username, password });
-        if (existinguser) {
-            const accessToken = jwt.sign({ id: existinguser._id }, JWT_ACCESS_PASSWORD, { expiresIn: '1m' });
-            const refreshToken = jwt.sign({ id: existinguser._id }, JWT_REFRESH_PASSWORD, { expiresIn: '5m' });
-            // res.cookie('accessToken', accessToken, { maxAge: 1 * 60 * 1000 })
+app.post('/api/v1/signin', async (req, res) => {
+    const username = req.body.username;
+    const password = req.body.password;
+    const existingUser = await User.findOne({ username: username, password: password });
+    if (existingUser) {
+        const accessToken = jwt.sign({ id: existingUser._id }, JWT_ACCESS_PASSWORD, { expiresIn: '1m' });
+        const refreshToken = jwt.sign({ id: existingUser._id }, JWT_REFRESH_PASSWORD, { expiresIn: '5m' });
+        res.cookie('refreshToken', refreshToken, { maxAge: 5 * 60 * 1000, httpOnly: true, secure: true, sameSite: 'strict' });
+        res.json({
+            message: 'User authenticated',
+            token: accessToken,
+            refresh: refreshToken
+        });
+    } else {
+        res.status(403).json({ message: 'Incorrect credentials' });
+    }
+});
 
+app.get('/api/v1/checkAuth', userMiddleware, (req, res) => {
+    res.status(200).json({ valid: true, message: "user authenticated" });
+});
 
-            res.cookie('refreshToken', refreshToken,
-                { maxAge: 5 * 60 * 1000, httpOnly: true, secure: true, sameSite: 'strict' })
+app.get('/api/v1/refresh', (req, res) => {
+    const refreshToken = req.cookies.refreshToken;
+    if (!refreshToken) {
+        res.status(401).json({ valid: false, message: "No refresh token provided" });
+    }
 
-            res.json({
-                message: "User authenticated",
-                token: accessToken,
-                refresh: refreshToken,
-            });
+    try {
+        const decoded = jwt.verify(refreshToken, JWT_REFRESH_PASSWORD) as JwtPayload;
+        const accessToken = jwt.sign(
+            {
+                id: decoded.id
+            },
+            JWT_ACCESS_PASSWORD,
+            { expiresIn: "1m" });
+
+        res.status(201).json({ accessToken: accessToken, message: "Access token generated" });
+    } catch (error) {
+        if (error.name === "TokenExpiredError") {
+            res.status(401).json({ valid: false, message: "Refresh token expired" });
         } else {
-            res.status(403).json({
-                message: "Incorrect credentials"
-            });
-        }
-    });
-
-app.post('api/v1/refresh',async(req,res)=>{
-    const refreshToken=req.cookies.refreshToken;
-    if(!refreshToken){
-        res.status(401).json({valid:false,message:"refresh Token expired"});
-    }else{
-        try {
-            const decoded=jwt.verify(refreshToken,JWT_REFRESH_PASSWORD);
-            if(decoded){
-                const accessToken=jwt.sign({id:(decoded as JwtPayload).id},JWT_ACCESS_PASSWORD,{expiresIn:'1m'});
-                res.status(201).json({accessToken:accessToken,message:"new access token generated"});
-            }else{
-                res.status(403).json({valid:false,message:"Invalid refresh token"});
-            }
-        } catch (error) {
-            throw new Error(error);
+            res.status(403).json({ valid: false, message: "Invalid refresh token" });
         }
     }
 })
 
-app.route('/api/v1/content')
-    .get(verifyuser, async (req, res) => {
-        const userId = req.userId;
-        const content = await Content.find({
-            userId: userId
-        }).populate("userId", "username");
-        res.status(200).json({
-            content
-        })
+
+
+app.get('/api/v1/content', userMiddleware, async (req, res) => {
+    const userId = req.userId;
+    const content = await Content.find({
+        userId: userId
+    }).populate("userId", "username");
+    res.status(200).json({
+        content
     })
-    .post(verifyuser, async (req, res) => {
-        const title = req.body.title;
-        const link = req.body.link;
-        const type = req.body.type;
-        await Content.create({
-            link,
-            type,
-            title,
+})
 
-            userId: req.userId,
-            tags: []
-        })
+app.post('/api/v1/content', userMiddleware, async (req, res) => {
+    const title = req.body.title;
+    const link = req.body.link;
+    const type = req.body.type;
+    await Content.create({
+        link,
+        type,
+        title,
+        userId: req.userId,
+        tags: []
+    })
 
-        res.json({
-            message: "Content added successfully "
-        })
-
-    });
-
-app.post('/api/v1/verifyJWT', verifyuser, async (req, res) => {
-    res.status(401).json({ valid: true, message: "user authorized" })
+    res.json({
+        message: "Content added successfully "
+    })
 });
 
-app.delete('/api/v1/:contentId', verifyuser, async (req, res) => {
+app.delete('/api/v1/:contentId', userMiddleware, async (req, res) => {
     const contentId = req.params.contentId;
     console.log(" deleting contentId", contentId);
     const deletedContent = await Content.findByIdAndDelete(contentId);
@@ -132,11 +126,10 @@ app.delete('/api/v1/:contentId', verifyuser, async (req, res) => {
     })
 });
 
-app.post('/api/v1/brain/share', verifyuser, async (req, res) => {
+app.post("/api/v1/brain/share", userMiddleware, async (req, res) => {
     const share = req.body.share;
     if (share) {
         const existingLink = await Link.findOne({
-
             userId: req.userId
         });
 
@@ -152,25 +145,20 @@ app.post('/api/v1/brain/share', verifyuser, async (req, res) => {
             userId: req.userId,
             hash: hash
         })
-
         res.json({
             hash
         })
     } else {
-        await Link.deleteOne({
-
-            userId: req.userId
-        });
-
+        await Link.deleteOne({ userId: req.userId });
         res.json({
             message: "Removed link"
         })
     }
 });
 
+
 app.get("/api/v1/brain/:shareLink", async (req, res) => {
     const hash = req.params.shareLink;
-
     const link = await Link.findOne({
         hash
     });
@@ -181,16 +169,13 @@ app.get("/api/v1/brain/:shareLink", async (req, res) => {
         })
         return;
     }
-    // userId
     const content = await Content.find({
         userId: link.userId
     })
-
     console.log(link);
     const user = await User.findOne({
         _id: link.userId
     })
-
     if (!user) {
         res.status(411).json({
             message: "user not found, error should ideally not happen"
@@ -204,10 +189,7 @@ app.get("/api/v1/brain/:shareLink", async (req, res) => {
 
 })
 
-
 connectDB();
-
-app.listen(port, () => {
+app.listen(3000, () => {
     console.log(`Server is running on port http://localhost:${port}`);
-
-});
+})
